renamed validator to conform  to the rest of the project
