+++ routes/owner.routes.js
@@ -2,9 +2,9 @@ const express = require('express');
 const router = express.Router();
 const { createOwner, getOwners, getOwnerById, getOwnerByEmail, deleteOwner} = require("../controllers/owner.controller");
 const { validateOwner } = require('../validators/ownerValidator');
-const { validateId } = require('../validators/idValidator');
+const { validateId } = require('../validators/id.validator');
 const validateResult = require('../middleware/resultValidator');
-const { validateEmail } = require('../validators/emailValidator');
+const { validateEmail } = require('../validators/email.validator');
 const checkUniqueField = require('../middleware/checkUniqueField');
 const { Owner } = require('../models');
